Stop the conversion upon error.
We can't just continue after an error, because we dont
know anymore which argument corresponds to which conversion
specifier.

For snprintf(), format errors result in undefined behavior, so
anything behavior is "acceptable". As snfmt is intended for
logging we don't want it to crash the program or abort(). So
just stop the conversion.